backend/requirements.txt
# Backend requirements for Dark Web Threat Intelligence system
# Lightweight version - heavy ML libs optional

# FastAPI & web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Kafka
kafka-python==2.0.2

# NLP & ML (lightweight - sklearn only)
scikit-learn==1.3.2
nltk==3.8.1
pandas==2.1.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional: Install separately if needed for advanced NLP
# transformers==4.35.2
# torch==2.1.1
